+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveLineService.java
@@ -28,8 +28,10 @@ import com.axelor.apps.stock.db.StockMoveLine;
 import com.axelor.apps.stock.db.TrackingNumber;
 import com.axelor.apps.stock.db.TrackingNumberConfiguration;
 import com.axelor.exception.AxelorException;
+import com.google.inject.persist.Transactional;
 import java.math.BigDecimal;
 import java.time.LocalDate;
+import java.util.LinkedHashMap;
 import java.util.List;
 
 public interface StockMoveLineService {
@@ -218,4 +220,8 @@ public interface StockMoveLineService {
    */
   public void setProductInfo(StockMove stockMove, StockMoveLine stockMoveLine, Company company)
       throws AxelorException;
+  
+  @Transactional
+  public void splitStockMoveLineByTrackingNumber(
+      StockMoveLine stockMoveLine, List<LinkedHashMap<String, Object>> trackingNumbers);
 }